Change location of ChessGame, ChessGameEvent and ChessGameListener to separate controller package, more logical place
