+++ src/controller/ChessGameEvent.java
@@ -1,4 +1,4 @@
-package model;
+package controller;
 
 import java.util.EventObject;
