+++ src/controller/ChessGameListener.java
@@ -1,4 +1,4 @@
-package model;
+package controller;
 
 import java.util.EventListener;
